# contracts/openapi/v1/provider.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAKEIN Provider API",
    "version": "1.0.0",
    "description": "Existing provider-owner and provider-branch API contract. Every operation requires Sanctum authentication plus the server-side provider account, branch, and menu permission checks identified by x-provider-permission."
  },
  "servers": [
    {
      "url": "https://api.takein.id/api",
      "description": "Production API ingress"
    }
  ],
  "paths": {
    "/provider/profile": {
      "get": {
        "operationId": "showProviderProfile",
        "summary": "Show the current provider profile",
        "x-laravel-route-name": "api.provider.profile.show",
        "x-authentication": "sanctum",
        "x-provider-permission": "profile",
        "security": [
          {
            "sanctum": []
          },
          {
            "sanctumSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProviderProfile",
        "summary": "Update the current provider profile",
        "x-laravel-route-name": "api.provider.profile.update",
        "x-authentication": "sanctum",
        "x-provider-permission": "profile",
        "security": [
          {
            "sanctum": []
          },
          {
            "sanctumSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Provider scope or permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonObject"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProviderProfileUpdateRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      },
      "sanctumSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "youyaku_session",
        "description": "Stateful Sanctum session cookie; deployment can configure its concrete name."
      }
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
      "ProviderProfileUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          },
          "phone_number": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "image": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary",
            "description": "Allowed media: JPEG, PNG, or WebP; maximum 2 MiB."
          }
        }
      },
...
    }
  }
}